requirements.txt
tensorflow-gpu
yacs
tqdm
opencv-python
